+++ addons/fray/lib/state_machine/state.gd
@@ -1,33 +1,17 @@
 extends Resource
-## docstring
+## Base state class
 
-#signals
-
-#enums
-
-#constants
-
-#preloaded scripts and scenes
-
-#exported variables
-
-var position: Vector2
-
-#private variables
-
-#onready variables
+func _enter_impl(arg = null) -> void:
+    pass
 
 
-#optional built-in virtual _init method
+func _update_impl(_delta: float) -> void:
+    pass
 
-#built-in virtual _ready method
 
-#remaining built-in virtual methods
+func _physics_update_impl(_delta: float) -> void:
+    pass
 
-#public methods
 
-#private methods
-
-#signal methods
-
-#inner classes
+func _exit_impl() -> void:
+    pass